Experiment: board with most words
Idea: how many unique Boggle boards are there, ignoring boards that are
rotationally identical and ignoring swapping dice in such a way that
they result in the same board? Some prior art on this point here:
https://www.danvk.org/wp/2007-08-02/how-many-boggle-boards-are-there/index.html

[That series](https://www.danvk.org/wp/2014-01-25/what-up-with-boggle/)
is focused on Boggle as an optimization problem, e.g. "what is the board
with the most words." My experiments are a little more geared towards
probabilistic questions: what does a typical Boggle word look like?
Which words are likely to occur? And so on.